opentelemetry-sdk/test-requirements.txt
asgiref==3.7.2
flaky==3.7.0
importlib-metadata==6.11.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psutil==7.2.2
py-cpuinfo==9.0.0
pytest==7.4.4
pyyaml==6.0.2
tomli==2.0.1
typing_extensions==4.10.0
wrapt==1.16.0
zipp==3.19.2
-e opentelemetry-api
-e tests/opentelemetry-test-utils
-e opentelemetry-semantic-conventions
-e opentelemetry-sdk
